cleaning up code examples
