+++ docs/components/Code.tsx
@@ -1,55 +1,13 @@
 import { usePathname } from 'next/navigation';
-import copy from 'clipboard-copy';
-import { createPreviewUrl, createUrl } from 'playroom/utils';
-import { Highlight, Prism } from 'prism-react-renderer';
-import React, {
-	Fragment,
-	type KeyboardEvent,
-	type ReactNode,
-	useCallback,
-	useContext,
-	useEffect,
-	useRef,
-	useState,
-} from 'react';
-import { LiveContext, LiveEditor, LivePreview, LiveProvider } from 'react-live';
-import { ExternalLinkCallout } from '@ag.ds-next/react/a11y';
-import { Box } from '@ag.ds-next/react/box';
-import { Button, ButtonLink } from '@ag.ds-next/react/button';
-import { CardHeader } from '@ag.ds-next/react/card';
-import {
-	globalPalette,
-	mapSpacing,
-	packs,
-	tokens,
-	useId,
-	useToggleState,
-} from '@ag.ds-next/react/core';
-import { ChevronRightIcon } from '@ag.ds-next/react/icon';
-import { Flex } from '@ag.ds-next/react/flex';
-import { Heading } from '@ag.ds-next/react/heading';
-import {
-	ChevronDownIcon,
-	ChevronUpIcon,
-	CopyIcon,
-} from '@ag.ds-next/react/icon';
-import {
-	proseBlockClassname,
-	unsetProseStylesClassname,
-} from '@ag.ds-next/react/prose';
-import { TextLink } from '@ag.ds-next/react/text-link';
+import React, { Fragment, type ReactNode, useState } from 'react';
+import { LiveProvider } from 'react-live';
 import { withBasePath } from '../lib/img';
-import useGetScrollbarWidth from '../lib/hooks/useGetScrollbarWidth';
 import * as designSystemComponents from './designSystemComponents';
-import { prismTheme } from './prism-theme';
-
-// Find multi-line comments at start `/** ... */`
-const multiLineCommentRegex = /^\/\*[\s\S]*?\*\//gi;
-
-// Add support for diff language support
-// https://github.com/FormidableLabs/prism-react-renderer#custom-language-support
-(typeof global !== 'undefined' ? global : window).Prism = Prism;
-require('prismjs/components/prism-diff');
+import { createTitleFromPathname } from './code/utils';
+import { ResponsivePreviewLink } from './code/ResponsivePreviewLink';
+import { StaticCode } from './code/StaticCode';
+import { LiveCode } from './code/LiveCode';
+import { PagePreviewIFrame } from './code/PagePreviewIFrame';
 
 export const PlaceholderImage = () => (
 	<img
@@ -80,417 +38,6 @@ export const PlaceholderPictogram = () => (
 		<path d="M43.0003 6L43 17C43 18.1045 43.8955 19 45 19H54" />
 	</svg>
 );
-
-function CodeHeading({
-	exampleContentHeadingType,
-	children,
-}: {
-	exampleContentHeadingType?: 'h2' | 'h3' | 'h4';
-	children: React.ReactNode;
-}) {
-	return (
-		<CardHeader>
-			<Heading type={exampleContentHeadingType}>{children}</Heading>
-		</CardHeader>
-	);
-}
-
-// Checks if the code requires React support or is JSX
-function checkAndModifyCode(liveCode: string) {
-	// Wrap `/* ... */` comments with brackets `{ .. }`
-	let code = liveCode.replaceAll(multiLineCommentRegex, '{$&}');
-	// No formatting required for JSX only
-	if (code.startsWith('<') || code.endsWith('>')) return code;
-
-	// Remove `;` from the end of `() => {};`
-	if (code.endsWith(';')) {
-		code = code.slice(0, -1);
-	}
-
-	const formattedCode = code.split('\n').join('\n    ');
-	return `<Render>\n    {${formattedCode}}\n</Render>`;
-}
-
-const removePath = [
-	'/components/',
-	'/content/',
-	'/foundations/',
-	'/guides/',
-	'/patterns/',
-	'/templates/',
-];
-
-function createTitleFromPathname(pathname: string) {
-	// Remove known path prefixing
-	const title = removePath.reduce((acc, str) => {
-		if (acc.startsWith(str)) {
-			return acc.replace(str, '');
-		}
-		return acc;
-	}, pathname);
-
-	return title
-		.split('/') // Nested paths
-		.map((str) => {
-			// Remove `-` from file names
-			const newStr = str.replaceAll('-', ' ');
-
-			// Capitalise first letter
-			return newStr.charAt(0).toUpperCase() + newStr.slice(1);
-		})
-		.join(': ');
-}
-
-// Query param keys used in responsive preview page
-export const responsivePreviewQueryKeys = {
-	disablePadding: 'disable-padding',
-	frameSrc: 'frame-src',
-	playroomSrc: 'playroom-code',
-	returnLink: 'return',
-	title: 'title',
-};
-
-export function ResponsivePreviewLink({
-	children,
-	code,
-	disablePadding = false,
-	frameAddress,
-	standalone = false,
-	title,
-}: {
-	children: React.ReactNode;
-	code?: string;
-	disablePadding?: boolean;
-	frameAddress?: string;
-	standalone?: boolean;
-	title: string;
-}) {
-	const urlParams = new URLSearchParams();
-
-	if (title) urlParams.append(responsivePreviewQueryKeys.title, title);
-
-	if (frameAddress)
-		urlParams.append(responsivePreviewQueryKeys.frameSrc, frameAddress);
-	if (code) {
-		const playroomPreviewUrl = createPreviewUrl({
-			baseUrl: process.env.NEXT_PUBLIC_PLAYROOM_URL,
-			code: checkAndModifyCode(code),
-		});
-		urlParams.append(
-			responsivePreviewQueryKeys.playroomSrc,
-			playroomPreviewUrl
-		);
-	}
-
-	if (disablePadding)
-		urlParams.append(responsivePreviewQueryKeys.disablePadding, 'true');
-
-	const pathname = usePathname();
-	urlParams.append(responsivePreviewQueryKeys.returnLink, pathname);
-
-	const href = `/preview/responsive?${urlParams.toString()}`;
-
-	// As link without code
-	if (standalone) {
-		return (
-			<Box css={{ marginTop: '1.5rem' }}>
-				<TextLink href={href}>
-					<Heading
-						alignItems="center"
-						as="span"
-						color="inherit"
-						gap={0.25}
-						type="h3"
-					>
-						{children}
-					</Heading>
-					<ChevronRightIcon css={{ verticalAlign: 'text-bottom' }} />
-				</TextLink>
-			</Box>
-		);
-	}
-
-	// Inline with code actions
-	return (
-		<ButtonLink href={href} size="sm" variant="tertiary">
-			{children}
-		</ButtonLink>
-	);
-}
-
-const PreviewActionContainer = ({
-	children,
-}: {
-	children: React.ReactNode;
-}) => (
-	<Flex
-		borderColor="muted"
-		borderTop
-		css={packs.print.hidden}
-		flexWrap="wrap"
-		gap={0.5}
-		padding={0.5}
-	>
-		{children}
-	</Flex>
-);
-
-const CopyCodeButton = ({ code }: { code: string }) => {
-	const copyLiveCode = useCallback(() => {
-		copy(code);
-	}, [code]);
-
-	return (
-		<Button
-			iconAfter={CopyIcon}
-			onClick={copyLiveCode}
-			size="sm"
-			variant="tertiary"
-		>
-			Copy code
-		</Button>
-	);
-};
-
-const OpenInPlayroomButton = ({ playroomUrl }: { playroomUrl: string }) => (
-	<ButtonLink href={playroomUrl} size="sm" variant="tertiary">
-		Open in Playroom
-		<ExternalLinkCallout />
-	</ButtonLink>
-);
-
-const ResponsivePreviewButton = ({
-	code,
-	disablePadding,
-	title,
-}: {
-	code: string;
-	disablePadding?: boolean;
-	title: string;
-}) => (
-	<ResponsivePreviewLink
-		code={code}
-		disablePadding={disablePadding}
-		title={title}
-	>
-		Preview responsive component
-	</ResponsivePreviewLink>
-);
-
-function LiveCode({
-	enableProse = false,
-	exampleContentHeading,
-	exampleContentHeadingType,
-	responsivePreviewHeading = 'Responsive preview',
-	showCode = false,
-	disablePadding,
-}: {
-	enableProse?: boolean;
-	exampleContentHeading?: string;
-	exampleContentHeadingType?: 'h2' | 'h3' | 'h4';
-	responsivePreviewHeading?: string;
-	showCode?: boolean;
-	disablePadding?: boolean;
-}) {
-	const liveEditorRef = useRef<HTMLDivElement>(null);
-	const liveCodeToggleButton = useRef<HTMLButtonElement>(null);
-	const live = useContext(LiveContext);
-
-	const liveOnChange = live.onChange;
-	const [localCopy, setLocalCopy] = useState<string>(live.code);
-	const [isCodeVisible, toggleIsCodeVisible] = useToggleState(
-		showCode,
-		!showCode
-	);
-
-	const handleChange = useCallback(
-		(code: string) => {
-			liveOnChange(code);
-			setLocalCopy(code);
-		},
-		[liveOnChange]
-	);
-
-	const codeUrl = useCallback(() => checkAndModifyCode(live.code), [live.code]);
-
-	const playroomUrl = createUrl({
-		baseUrl: process.env.NEXT_PUBLIC_PLAYROOM_URL,
-		code: codeUrl(),
-	});
-
-	const id = useId();
-	const codeId = `live-code-${id}`;
-
-	const onLiveEditorContainerKeyDown = useCallback(
-		(event: KeyboardEvent<HTMLDivElement>) => {
-			if (event.code === 'Escape') {
-				toggleIsCodeVisible();
-				liveCodeToggleButton.current?.focus();
-			}
-		},
-		[toggleIsCodeVisible]
-	);
-
-	// LiveEditor doesn't support aria-label, so we have to do it the DOM way
-	useEffect(() => {
-		const pre = liveEditorRef.current?.querySelector('pre');
-		if (pre) {
-			pre.ariaLabel = `Live code editor ${id}`;
-			// eslint-disable-next-line @typescript-eslint/ban-ts-comment
-			// @ts-ignore Property 'role' does not exist on type 'HTMLPreElement'
-			pre.role = 'region';
-		}
-	}, [id]);
-
-	// Using `Box` here instead of Code snippets with popovers (date picker, combobox, dropdown menu etc) need overflow
-	return (
-		<Box border borderColor="muted" className={proseBlockClassname} rounded>
-			{exampleContentHeadingType && (
-				<CodeHeading exampleContentHeadingType={exampleContentHeadingType}>
-					{exampleContentHeading}
-				</CodeHeading>
-			)}
-			<LivePreview
-				aria-label={`Rendered code snippet example ${id}`}
-				// Prevents prose styles from being inherited in live code examples (except for the prose example)
-				className={enableProse ? undefined : unsetProseStylesClassname}
-				css={{
-					// The mdx codeblock transform wraps the code component in a pre which
-					// applies some weirdness here. This resets back to normal things
-					whiteSpace: 'normal', // other wise text content will not wrap and long lines can break the layout
-					fontFamily: tokens.font.body, // because pre applies gets monospace font.
-					padding: mapSpacing(1.5),
-				}}
-				role="region"
-			/>
-			<PreviewActionContainer>
-				<Button
-					aria-controls={codeId}
-					aria-expanded={isCodeVisible}
-					iconAfter={isCodeVisible ? ChevronUpIcon : ChevronDownIcon}
-					onClick={toggleIsCodeVisible}
-					ref={liveCodeToggleButton}
-					size="sm"
-					variant="tertiary"
-				>
-					{isCodeVisible ? 'Hide live code' : 'Show live code'}
-				</Button>
-				<CopyCodeButton code={localCopy} />
-				<OpenInPlayroomButton playroomUrl={playroomUrl} />
-				<ResponsivePreviewButton
-					code={live.code}
-					disablePadding={disablePadding}
-					title={responsivePreviewHeading}
-				/>
-			</PreviewActionContainer>
-			<Box
-				css={packs.print.visible}
-				display={isCodeVisible ? 'block' : 'none'}
-				id={codeId}
-				onKeyDown={onLiveEditorContainerKeyDown}
-				palette="dark"
-				ref={liveEditorRef}
-			>
-				<LiveEditor
-					code={live.code}
-					css={{
-						'&:focus-within': packs.outline,
-						'textarea, pre': {
-							padding: `${mapSpacing(1.5)} !important`,
-							tabSize: `4 !important`,
-						},
-						'& ::selection': {
-							color: globalPalette.darkBackgroundBody,
-							backgroundColor: globalPalette.darkForegroundAction,
-						},
-					}}
-					disabled={live.disabled}
-					language={live.language}
-					onChange={handleChange}
-					tabMode="focus"
-					theme={prismTheme}
-				/>
-			</Box>
-			{live.error ? (
-				<Box
-					background="shade"
-					color="error"
-					fontFamily="monospace"
-					fontSize="xs"
-					padding={1}
-				>
-					{live.error}
-				</Box>
-			) : null}
-		</Box>
-	);
-}
-
-const StaticCode = ({
-	code,
-	language = '', // By default render as plain text (ie. no language)
-	inline = false,
-}: {
-	code: string;
-	language?: string;
-	inline?: boolean;
-}) => {
-	return (
-		<Box
-			border
-			borderColor="muted"
-			css={{
-				marginTop: inline ? 0 : mapSpacing(1.5),
-
-				pre: {
-					padding: `${mapSpacing(1.5)} !important`,
-					overflowX: 'auto',
-				},
-
-				'& ::selection': {
-					color: globalPalette.darkBackgroundBody,
-					backgroundColor: globalPalette.darkForegroundAction,
-				},
-			}}
-			rounded
-		>
-			<Box dark>
-				<Highlight code={code} language={language} theme={prismTheme}>
-					{({ className, style, tokens, getLineProps, getTokenProps }) => (
-						<pre
-							className={[className, unsetProseStylesClassname].join(' ')}
-							style={style}
-						>
-							<code>
-								{tokens.map((line, lineKey) => (
-									<div key={lineKey} {...getLineProps({ line, key: lineKey })}>
-										{line.map((token, tokenKey) => (
-											<span
-												key={tokenKey}
-												{...getTokenProps({ token, key: tokenKey })}
-											/>
-										))}
-									</div>
-								))}
-							</code>
-						</pre>
-					)}
-				</Highlight>
-			</Box>
-			<Flex padding={0.5}>
-				<Button
-					aria-label="Copy code snippet to clipboard"
-					iconAfter={CopyIcon}
-					onClick={() => copy(code)}
-					size="sm"
-					variant="tertiary"
-				>
-					Copy code
-				</Button>
-			</Flex>
-		</Box>
-	);
-};
 
 const LIVE_SCOPE = {
 	...designSystemComponents,
@@ -550,7 +97,7 @@ export function Code({
 
 	if (iFrame) {
 		return (
-			<PagePreviewFrame
+			<PagePreviewIFrame
 				code={childrenAsString}
 				language={language}
 				showCode={showCode}
@@ -579,85 +126,4 @@ export function Code({
 	}
 
 	return <StaticCode code={childrenAsString} language={language} />;
-}
-
-function PagePreviewFrame({
-	code,
-	language,
-	showCode = false,
-	title,
-	disablePadding,
-}: {
-	code: string;
-	language?: string;
-	showCode?: boolean;
-	title: string;
-	disablePadding?: boolean;
-}) {
-	const scrollbarWidth = useGetScrollbarWidth();
-	const [isCodeVisible, toggleIsCodeVisible] = useToggleState(
-		showCode,
-		!showCode
-	);
-
-	const id = useId();
-	const codeId = `live-code-${id}`;
-
-	const playroomUrl = createUrl({
-		baseUrl: process.env.NEXT_PUBLIC_PLAYROOM_URL,
-		code: checkAndModifyCode(code),
-	});
-	const playroomPreviewUrl = createPreviewUrl({
-		baseUrl: process.env.NEXT_PUBLIC_PLAYROOM_URL,
-		code: checkAndModifyCode(code),
-	});
-
-	return (
-		<Box border borderColor="muted">
-			<div
-				css={{
-					margin: '1rem',
-					marginRight: `calc(1rem - ${scrollbarWidth}px)`,
-					marginBottom: 0,
-				}}
-			>
-				<iframe
-					css={{
-						width: '100%',
-						height: '36rem',
-						border: 0,
-					}}
-					src={playroomPreviewUrl}
-					title={title}
-				></iframe>
-			</div>
-			<PreviewActionContainer>
-				<Button
-					aria-controls={codeId}
-					aria-expanded={isCodeVisible}
-					iconAfter={isCodeVisible ? ChevronUpIcon : ChevronDownIcon}
-					onClick={toggleIsCodeVisible}
-					size="sm"
-					variant="tertiary"
-				>
-					{isCodeVisible ? 'Hide code' : 'Show code'}
-				</Button>
-				<CopyCodeButton code={code} />
-				<OpenInPlayroomButton playroomUrl={playroomUrl} />
-				<ResponsivePreviewButton
-					code={code}
-					disablePadding={disablePadding}
-					title={title}
-				/>
-			</PreviewActionContainer>
-			<Box
-				css={packs.print.visible}
-				display={isCodeVisible ? 'block' : 'none'}
-				id={codeId}
-				palette="dark"
-			>
-				<StaticCode code={code} inline language={language} />
-			</Box>
-		</Box>
-	);
 }
