+++ packages/react/src/table/docs/overview.mdx
@@ -8,7 +8,14 @@ relatedComponents: ['summary-list']
 relatedPatterns: ['selectable-table-with-batch-actions', 'search-filters']
 ---
 
-```jsx responsivePreviewHeading="Preview viewing a record in a table"
+<ResponsivePreview
+	disablePadding
+	href="/yourgov"
+	label="Preview using href"
+	title="Yourgov preview"
+/>
+
+```jsx responsivePreviewHeading="Preview using code"
 <TableWrapper>
 	<Table>
 		<TableCaption>Applications (with actions dropdown)</TableCaption>
